requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.7
oauthlib==2.1.0
python-dotenv==0.8.2
python-twitter==3.4.2
requests==2.19.1
requests-oauthlib==1.0.0
thesaurus==0.2.2
urllib3==1.23
